[Feature-MakeClub] Move make club file to club package
