+++ features/feature-makeclub/src/main/java/com/example/feature_makeclub/presentation/ui_component/make_club/ClubNameView.kt
@@ -1,4 +1,4 @@
-package com.example.feature_makeclub.presentation.ui_component
+package com.example.feature_makeclub.presentation.ui_component.make_club
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
